+++ cryptofeed/raw_data_collection.py
@@ -90,7 +90,7 @@ async def _playback(feed: str, filenames: list, callbacks: dict, config: str):
     else:
         for ctype in callbacks.keys():
             callbacks[ctype] = [callbacks[ctype], functools.partial(internal_cb, cb_type=ctype)]
-    feed = EXCHANGE_MAP[feed](config=config, subscription=sub, callbacks=callbacks)
+    feed = EXCHANGE_MAP[feed](candle_closed_only=False, config=config, subscription=sub, callbacks=callbacks)
 
     exchange_sub = {}
     for chan in ws.subscription: