+++ CHANGES.md
@@ -1,5 +1,19 @@
 ## Changelog
 
+### 2.2.2
+ * Bugfix: OKX filled amount being reported incorrectly in OrderInfo
+ * Bugfix: Tweak QuestDB coulmn types and layout
+ * Bugfix: Fix Bybit Private Channel connections / subscriptions
+ * Bugfix: Return client order id in OrderInfo object returned by Coinbase
+ * Feature: Add Order type
+ * Feature: Add support for closed candles only in Bybit
+ * Update: Kraken Futures new instrument type: Perpetual Linear Multi-collateral Futures
+ * New Exchange: Bitget
+ * New Exchange: Independent Reserve
+ * Feature: Add perpetuals to Bitget
+ * Update: Add indicator in symbol info if instrument is a qunto
+ * Feature: Configuration option to allow invalid symbols
+
 ### 2.2.1 (2022-02-27)
  * Feature: Support for order info stream on BitMEX
  * Bugfix: Datetime/Timestamp conversion fixes